requirements.txt
Django==3.0.5
djangorestframework_jwt==1.11.0
django_jsonfield==1.4.1
djangorestframework==3.12.4
